[refactor] CommandBufferManager 제거 후 SSDManager로 병합
